requirements.txt
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
Brotli==1.1.0
cachetools==6.2.4
certifi==2023.11.17
cffi==2.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
converter==1.0.0
converters==0.0.1
cryptography==46.0.5
css-html-js-minify==2.5.5
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
distlib==0.3.8
distro==1.9.0
faiss-cpu==1.12.0
ffmpeg==1.4
filelock==3.13.1
Flask==3.0.2
fonttools==4.53.0
fpdf==1.7.2
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.0
google-genai==1.66.0
h11==0.16.0
htmlmin==0.1.12
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.6
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.5.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.5
lxml==5.1.0
MarkupSafe==2.1.5
matplotlib==3.9.0
more-itertools==10.8.0
mpmath==1.3.0
mutagen==1.47.0
narwhals==2.18.0
networkx==3.5
numpy==1.26.3
outcome==1.3.0.post0
packaging==24.1
pandas==2.1.4
pillow==10.3.0
platformdirs==4.2.0
premailer==3.10.0
protobuf==6.33.5
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycountry==23.12.11
pycparser==2.21
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyparsing==3.1.2
pyperclip==1.8.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.2.2
pytz==2023.3.post1
PyYAML==6.0.2
referencing==0.37.0
regex==2025.9.1
reportlib==3.4.0
requests==2.32.3
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
selenium==4.16.0
sentence-transformers==5.1.0
setuptools==69.0.3
six==1.16.0
smmap==5.0.3
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
streamlink==6.5.1
streamlit==1.55.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.0
toml==0.10.2
torch==2.8.0
tornado==6.5.5
tqdm==4.67.1
transformers==4.56.1
trio==0.24.0
trio-websocket==0.11.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2023.4
urllib3==2.1.0
virtualenv==20.25.0
watchdog==6.0.0
websocket-client==1.7.0
websockets==13.0.1
Werkzeug==3.0.1
wordcloud==1.9.3
wsproto==1.2.0
youtube-dl==2021.12.17
yt-dlp==2024.8.6
